+++ eras/alonzo/impl/src/Cardano/Ledger/Alonzo/BlockBody.hs
@@ -1,0 +1,5 @@
+module Cardano.Ledger.Alonzo.BlockBody (
+  AlonzoBlockBody (AlonzoBlockBody),
+) where
+
+import Cardano.Ledger.Alonzo.BlockBody.Internal