+++ eras/shelley/impl/src/Cardano/Ledger/Shelley/BlockBody.hs
@@ -1,0 +1,11 @@
+module Cardano.Ledger.Shelley.BlockBody (
+  ShelleyBlockBody (ShelleyBlockBody),
+  auxDataSeqDecoder,
+  hashShelleySegWits,
+  bBodySize,
+  slotToNonce,
+  incrBlocks,
+  coreAuxDataBytes,
+) where
+
+import Cardano.Ledger.Shelley.BlockBody.Internal