+++ eras/shelley/test-suite/test/Test/Cardano/Ledger/Shelley/Serialisation/Golden/Encoding.hs
@@ -77,7 +77,7 @@ import Cardano.Ledger.PoolParams (
  )
 import Cardano.Ledger.Shelley (ShelleyEra)
 import Cardano.Ledger.Shelley.API (MultiSig)
-import Cardano.Ledger.Shelley.BlockChain (ShelleyTxSeq (..), bbHash)
+import Cardano.Ledger.Shelley.BlockBody (ShelleyBlockBody (..))
 import Cardano.Ledger.Shelley.Core
 import Cardano.Ledger.Shelley.PParams (
   ProposedPPUpdates (..),
@@ -145,7 +145,7 @@ import Test.Cardano.Ledger.Api.Examples.Consensus.Shelley as Ex (
 import Test.Cardano.Ledger.Binary.TreeDiff (CBORBytes (CBORBytes), ansiDocToString, diffExpr)
 import Test.Cardano.Ledger.Core.KeyPair (KeyPair (..), mkWitnessVKey, sKey, vKey)
 import Test.Cardano.Ledger.Shelley.ConcreteCryptoTypes (C, MockCrypto)
-import Test.Cardano.Ledger.Shelley.Generator.Core (KESKeyPair (..), PreAlonzo, VRFKeyPair (..))
+import Test.Cardano.Ledger.Shelley.Generator.Core (KESKeyPair (..), VRFKeyPair (..))
 import Test.Cardano.Ledger.Shelley.Generator.EraGen (genesisId)
 import Test.Cardano.Ledger.Shelley.Serialisation.GoldenUtils (
   ToTokens (..),
@@ -306,8 +306,7 @@ testHeaderHash =
 
 testBHB ::
   forall era.
-  ( EraTx era
-  , PreAlonzo era
+  ( EraBlockBody era
   , Tx era ~ ShelleyTx era
   ) =>
   BHBody MockCrypto
@@ -333,7 +332,7 @@ testBHB =
           )
           (vrfSignKey testVRF)
     , bsize = 0
-    , bhash = bbHash @era $ ShelleyTxSeq @era StrictSeq.empty
+    , bhash = hashBlockBody $ mkBasicBlockBody & txSeqBlockBodyL @era .~ StrictSeq.empty
     , bheaderOCert =
         OCert
           (kesVerKey testKESKeys)
@@ -348,8 +347,7 @@ testBHB =
 
 testBHBSigTokens ::
   forall era.
-  ( EraTx era
-  , PreAlonzo era
+  ( EraBlockBody era
   , Tx era ~ ShelleyTx era
   ) =>
   Tokens ->
@@ -900,7 +898,7 @@ tests =
               (vrfSignKey testVRF)
           size = 0
           blockNo = BlockNo 44
-          bbhash = bbHash @C $ ShelleyTxSeq StrictSeq.empty
+          bbhash = hashBlockBody @C $ ShelleyBlockBody StrictSeq.empty
           ocert :: OCert MockCrypto
           ocert =
             OCert
@@ -978,7 +976,7 @@ tests =
       let sig :: (SignedKES (KES MockCrypto) (BHBody MockCrypto))
           sig = unsoundPureSignedKES () 0 (testBHB @C) (kesSignKey testKESKeys)
           bh = BHeader (testBHB @C) sig
-          txns = ShelleyTxSeq StrictSeq.Empty
+          txns = ShelleyBlockBody StrictSeq.Empty
        in checkEncodingCBORAnnotated
             shelleyProtVer
             "empty_block"
@@ -1039,7 +1037,7 @@ tests =
             mkBasicTx txb5
               & witsTxL @C .~ (mkBasicTxWits @C & addrTxWitsL .~ ws & scriptTxWitsL .~ ss)
               & auxDataTxL @C .~ SJust tx5MD
-          txns = ShelleyTxSeq $ StrictSeq.fromList [tx1, tx2, tx3, tx4, tx5]
+          txns = ShelleyBlockBody $ StrictSeq.fromList [tx1, tx2, tx3, tx4, tx5]
        in checkEncodingCBORAnnotated
             shelleyProtVer
             "rich_block"